created animations folder and BirdAnimation module, added top position in the setBirdPosition function
