+++ src/js/modules/gameAlgorithm.mjs
@@ -4,7 +4,7 @@ import { MusicIsActive, playDefaultSoundTrack, renderPlayMusicButtons, setDefaul
 import { TOP_BAR_CONTAINER, memoryDeck } from './m-themes/addCards.mjs';
 import { DECK_CONTAINER } from './m-themes/deckStyles.mjs';
 import { difficulty } from './m-themes/themesDifficulty.mjs';
-import { resetBirdAnimation, startBirdAnimation } from './menuActions.mjs';
+import { resetBirdAnimation, startBirdAnimation } from './animations/forest_theme/BirdAnimation.mjs';
 import { fillRandomThemes } from './randomThemes/fillRandomThemes.mjs';
 
 const SCOREBOARD = document.getElementById('scorePoints');
@@ -176,7 +176,7 @@ function endGame() {
   hideElements(DECK_CONTAINER);
   hideElements(topBarContainerIngameElements);
   fillRandomThemes();
-  startBirdAnimation();
+  timeoutFunctionForTwoSeconds(startBirdAnimation);
   revealElements(GAME_MENU);
 }
