+++ src/js/modules/menuActions.mjs
@@ -13,8 +13,6 @@ const SETTINGS_BUTTONS = document.querySelectorAll('.settingButton');
 const CLOSE_MENU_BUTTONS = document.querySelectorAll('.close-icon');
 const TOP_BAR_CONTAINER = document.querySelector('.top_bar_container');
 
-let birdAnimationInterval;
-
 function openMenu() {
   playClickSoundEffect();
   let buttonDataset = this.dataset.setting;
@@ -76,61 +74,4 @@ MENU_SETTINGS_OPTIONS.forEach(SETTING_BUTTON => {
   SETTING_BUTTON.addEventListener('mouseleave', stopHoverSoundEffect);
 });
 
-// Bird Animation
-const BIRD_CONTAINER = document.querySelector('.bird_animated_gif_container');
-
-function startBirdAnimation() {
-  BIRD_CONTAINER.classList.add('flying_animation');
-  switchBirdState();
-  setBirdAnimationInterval();
-}
-
-function setBirdAnimationInterval() {
-  birdAnimationInterval = setInterval(switchBirdState, 20000);
-}
-
-function setBirdPosition() {
-  // get title (h1) right position
-  let titlePosition = document
-    .querySelector('#game-menu-title')
-    .getBoundingClientRect()
-    .right.toFixed();
-
-  // set bird position through CSS variable at :root
-  let root = document.querySelector(':root');
-  root.style.setProperty(
-    '--bird_left_position_animation',
-    titlePosition - 30 + 'px'
-  );
-
-  /* W3Schools helped me to code the setBirdPosition algorithm:
-    - https://www.w3schools.com/css/css3_variables_javascript.asp
-  */
-}
-
-function resetBirdAnimation() {
-  clearInterval(birdAnimationInterval);
-  BIRD_CONTAINER.classList.remove('flying_animation');
-}
-
-function switchBirdState() {
-  setTimeout(() => {
-    BIRD_CONTAINER.firstElementChild.src =
-      '../src/assets/images/templates/forest_theme/perched_bird.png';
-  }, 8000);
-
-  setTimeout(() => {
-    BIRD_CONTAINER.firstElementChild.src =
-      '../src/assets/gifs/single-bird-flying.gif';
-  }, 14000);
-}
-
-export {
-  SETTINGS_BUTTONS,
-  CLOSE_MENU_BUTTONS,
-  openMenu,
-  closeMenu,
-  startBirdAnimation,
-  setBirdPosition,
-  resetBirdAnimation
-};
+export { SETTINGS_BUTTONS, CLOSE_MENU_BUTTONS, openMenu, closeMenu };
